fix(backend): await async health check and move it to integration suite
