+++ backend/tests/unit/api/test_health.py
@@ -7,12 +7,3 @@ def test_root(client):
 
     assert resp.status_code == status.HTTP_200_OK
     assert data["message"] == "Skillora analytics backend is running"
-
-
-def test_health(client):
-    resp = client.get("/health")
-    data = resp.json()
-
-    assert resp.status_code == status.HTTP_200_OK
-    assert data["application"] == "ok"
-    assert data["db"] == "ok"
